Adapted Augst to CSM

+++ apps/app_demo_slproject/source/AppDemoGui.cpp
@@ -247,10 +247,10 @@ void AppDemoGui::build(SLProjectScene* s, SLSceneView* sv)
             }
 
             ImGui::Separator();
-            ImGui::Text("Parallel Jobs to follow: %llu",
-                        AppDemo::jobsToBeThreaded.size());
-            ImGui::Text("Sequential Jobs to follow: %llu",
-                        AppDemo::jobsToFollowInMain.size());
+            ImGui::Text("Parallel Jobs to follow: %u",
+                        (uint)AppDemo::jobsToBeThreaded.size());
+            ImGui::Text("Sequential Jobs to follow: %u",
+                        (uint)AppDemo::jobsToFollowInMain.size());
             ImGui::End();
             return;
         }
